Add some personalization options to the browser. (#35)
* Add Personalization JS File.

This is used to show or hide the elements in the settings menu related
to personalization in the browser.

* Update Default HTML for Personalization Options.

* Update CSS and Browser.js for Personalization Menu

* Fix missing personalization button.

This fixes the missing personalization button in the settings menu.

* Add Personalization.JS to Project File

* Add Local Storage API For Navigation Bar Color.

This uses the HTML 5 Local Storage API to store and retrieve the navigation bar color parameter. This is added so the settings are saved as they are set. That being said. This API will only store a max of 5mb.